Grupo 27 seller orders details

+++ back-end/src/database/migrations/20221027184359-created-sales.js
@@ -8,7 +8,7 @@ module.exports = {
         allowNull: false,
       },
       totalPrice: {
-        type: Sequelize.DECIMAL(9,2),
+        type: Sequelize.FLOAT(10,2),
         allowNull: false,
         field: 'total_price',
       },
